Implement non-recursive quicksort on SLL
The quicksort_non_recursive function has two arrays of node_t
pointers (i.e., node_t *). The two arrays are named `begin` and
`end`, and they are used to store the address of the beginning
and ending node in the list, respectively.

Similar to conventional quicksort on SLL, the
quicksort_non_recursive function selects the first node in the
list as the pivot. Then, the function splits the list into two
sublists, which are `left` and `right`. Every node with value
greater than that of the pivot is inserted into `right`;
otherwise, the node is inserted into `left`.

Concatenate `left`, `pivot`, and `right` after splitting the list.
Then, mark the beginning and ending node of both sublists with
`begin` and `end`, respectively. The addresses stored in `begin`
and `end` are shown as follows:
- begin[i]: the beginning node in `left`
- end[i]: the ending node in `left`
- begin[i + 1]: the beginning node in `right`
- end[i + 1]: the ending node in `right`.

If the length of `right` is greater than 1, sort `right` first;
otherwise, sort `left` first. If the length of `left` is less
than 2, concatenate `left` with the node on its left hand side.
Finally, if the level become 0 and the length of `left` is less
than 2, update the head of the list and end the function.

Note: In the worst-case scenario, which means that the list is
already sorted in ascending order, the quicksort_non_recursive
function can only sort lists of length less than `MAX_LEVELS`.